+++ grid2op/tests/BaseRedispTest.py
@@ -59,13 +59,15 @@ class BaseTestRedispatch(MakeBackend):
         # _parameters for the environment
         self.env_params = Parameters()
         self.env_params.ALLOW_DISPATCH_GEN_SWITCH_OFF = False
-        self.env = Environment(init_grid_path=os.path.join(self.path_matpower, self.case_file),
-                               backend=self.backend,
-                               chronics_handler=self.chronics_handler,
-                               parameters=self.env_params,
-                               names_chronics_to_backend=self.names_chronics_to_backend,
-                               actionClass=BaseAction,
-                               name="test_redisp_env1")
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore")
+            self.env = Environment(init_grid_path=os.path.join(self.path_matpower, self.case_file),
+                                   backend=self.backend,
+                                   chronics_handler=self.chronics_handler,
+                                   parameters=self.env_params,
+                                   names_chronics_to_backend=self.names_chronics_to_backend,
+                                   actionClass=BaseAction,
+                                   name="test_redisp_env1")
         self.array_double_dispatch = np.array([0.,  10.,  20.,   0., -30.])
         # self.array_double_dispatch = np.array([0.,  11.208119,  12.846733, 0., -24.054852])
         self.tol_one = self.env._tol_poly
@@ -322,13 +324,15 @@ class BaseTestRedispatchChangeNothingEnvironment(MakeBackend):
         # _parameters for the environment
         self.env_params = Parameters()
         self.env_params.ALLOW_DISPATCH_GEN_SWITCH_OFF = False
-        self.env = Environment(init_grid_path=os.path.join(self.path_matpower, self.case_file),
-                               backend=self.backend,
-                               chronics_handler=self.chronics_handler,
-                               parameters=self.env_params,
-                               names_chronics_to_backend=self.names_chronics_to_backend,
-                               actionClass=BaseAction,
-                               name="test_redisp_env2")
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore")
+            self.env = Environment(init_grid_path=os.path.join(self.path_matpower, self.case_file),
+                                   backend=self.backend,
+                                   chronics_handler=self.chronics_handler,
+                                   parameters=self.env_params,
+                                   names_chronics_to_backend=self.names_chronics_to_backend,
+                                   actionClass=BaseAction,
+                                   name="test_redisp_env2")
         self.tol_one = self.env._tol_poly
 
     def tearDown(self):